#33 wip split currency utils
